Working with front end features

+++ back-end/src/presentation/article/article.controller.ts
@@ -1,6 +1,6 @@
 
 import { Request, Response } from "express";
-import { ArticleService } from "../services/article.service";
+import { ArticleService } from "../services/article.service.js";
 
 
 export class ArticleController{
@@ -55,7 +55,10 @@ export class ArticleController{
 
     addComment = async (req: Request, res: Response) => {
         const name: string = req.params.name as string;
-        const {text, postedBy} = req.body;      
+        const {text, postedBy} = req.body;
+
+        const {uid} = (req as any).user;
+        console.log(uid);
 
         const article = this.articleService.getArticleByName(name);        
         if(!article){
@@ -70,7 +73,11 @@ export class ArticleController{
             return res.status(404).json("postedBy is required.");            
         }
 
-        const newComment = await this.articleService.addComment(name, text, postedBy);
+        if(!uid){
+            return res.status(404).json("userId is required.");            
+        }
+
+        const newComment = await this.articleService.addComment(name, text, postedBy, uid);
 
         if(!newComment)
         {
@@ -83,6 +90,23 @@ export class ArticleController{
 
         const name: string = req.params.name as string;
         const id: string = req.params.id as string;
+        const {uid} = (req as any).user;
+        const article = await this.articleService.getArticleByName(name);
+
+        if(!article)
+        {
+            return res.status(400).json("Article not found.");
+        }
+
+        const comment = article.comments?.find(f=>f.id === id);
+
+        if(!comment){
+            return res.status(400).json("Comment not found.");
+        }
+
+        if(comment.userId != uid){
+            return res.status(400).json("Comment doesn't belong you");
+        }
 
         if(!await this.articleService.deleteComment(name, id)){
             return res.status(400).json("Article not found.");
